Add files used to create surrogates
The surrogates created include those for Turbine 1 (upwind) and Turbine 2 (downwind) across 36 inflow conditions. The 36 inflow conditions are combinations of 8, 12, and 18 mps wind speeds, 5, 10, 15, and 20% turbulence, and 01, 0.2, and 0.3 wind shear. The interpolant was made from scipy 2D interpolation over the calculated L10 of Turbine 2 placed downwind of Turbine 1 at 624 locations. The locations included combinations of -3: 0.25: 3 crosswind of Turbine 1 and 3: 0.5: 15 downwind of Turbine 1.